[ibmpc] Better throttling of keyboard interrupts
If the host key repeat rate is higher than what a PC expects,
keys will be lost. As a workaround, whenever we are about to
trigger a keyboard interrupt, we check if the last interrupt
has been acknowleged yet. If not, we wait.

This patch also makes sure that we do not generate keyboard
interrupts for releasing a key that isn't pressed (the terminal
can generate such events).