[ibmpc] Patch the BIOS at runtime
With this patch there is no longer a need for a modified BIOS image.

If the PCE extension ROM is present at F000:0000 a jump to it is
automatically patched into the main BIOS at F000:FFF0

Calls to int 0x19 are intercepted and redirected to the PCE extension
ROM as they occur. The extension ROM then grabs int 0x13 before
trying to boot from the boot drive.