[68000] A hack to make System 6 boot
System 6 fails to boot with the custom floppy disk driver. It tries to
free some memory that the ROM driver presumably allocated but the custom
driver does not.

This issue should be fixed either in the custom floppy disk driver or
with a proper emulation of the floppy disk drive. For now, this is
an ugly workaround.