video: Update the clock when the status register is read
Since the clock function is not called every time the video clock is
updated, it's possible that the emulated program can read the status
register and determine that the vertical retrace has started before
the screen is actually updated. This is fixed by calling clock before
returning the status register.