macplus: Use VIA CB2 output for keyboard protocol
When the keyboard wants to send a response to the VIA it needs to
wait until the data line (CB2) goes high, which happens when CB2
is switched from output to input in the VIA.

The old code just delayed the keyboard response by 8192 clock cycles
and hoped for the best. With this patch the state of the data line
is actually checked.